feature/SCIHUB-288 fixed file validation
