+++ src/lib/packages/EditorContext/hooks/useUploadPlugin/useUploader.tsx
@@ -2,37 +2,13 @@ import { Schema, Node } from '@milkdown/prose/model';
 import { useCallback } from 'react';
 
 import { useTextEditorContext } from '../../../../components/TextEditorContext/useTextEditoContext';
-import { errorMessages } from '../../../../config/errorMessages';
+import { useFileValidation } from '../../../../hooks/useFileValidation';
 import { useNotification } from '../../../../hooks/useNotification';
-import { Matcher } from '../../../../utils/Matcher';
-
-const isFormatValid = (format: string, acceptedFiles: string[]) => {
-  const splittedFormat = format.split('/');
-  const { 0: type, [splittedFormat.length - 1]: extension } = splittedFormat;
-
-  if (acceptedFiles.includes('*')) {
-    return true;
-  }
-
-  return Matcher(type)
-    .match('image', () => {
-      if (acceptedFiles.includes('image/*')) {
-        return true;
-      }
-      return acceptedFiles.includes(`.${extension}`);
-    })
-    .match('application', () => {
-      if (acceptedFiles.includes('application/*')) {
-        return true;
-      }
-      return acceptedFiles.includes(`.${extension}`);
-    })
-    .getOrElse(() => false);
-};
 
 export const useUploader = () => {
   const { onErrorNotification } = useNotification();
-  const { onFileUpload, acceptedFormats } = useTextEditorContext();
+  const { onFileUpload } = useTextEditorContext();
+  const { isFileValid } = useFileValidation();
 
   const uploader = useCallback(
     async (files: FileList, schema: Schema): Promise<Node[]> => {
@@ -41,14 +17,10 @@ export const useUploader = () => {
 
         for (let i = 0; i < files.length; i++) {
           const file = files.item(i);
-          if (!file) {
+          if (!isFileValid(file)) {
             continue;
           }
-          if (!isFormatValid(file.type, acceptedFormats)) {
-            onErrorNotification(errorMessages.image.format);
-            continue;
-          }
-          images.push(file);
+          images.push(file as File);
         }
 
         return await Promise.all(
@@ -66,7 +38,7 @@ export const useUploader = () => {
         return [];
       }
     },
-    [onFileUpload, onErrorNotification, acceptedFormats]
+    [onFileUpload, onErrorNotification, isFileValid]
   );
 
   return uploader;
